requirements.txt
streamlit==1.32.0
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
numpy==1.26.4
transformers==4.39.3
huggingface_hub==0.23.1 